+++ lib/data/state/comment_for_campaign_notifier.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class CommentForCampaignNotifier extends StateNotifier<String?> {
+  CommentForCampaignNotifier() : super(null);
+
+  void setComment(String comment) {
+    state = comment;
+  }
+}